[Android] Fix for mono build linker error
