+++ browser/android/BUILD.gn
@@ -6,7 +6,7 @@ source_set("android_browser_process") {
 
   sources = [
     "//brave/browser/brave_ads/android/brave_ads_native_helper.cc",
-    "brave_feature_list.cc",
+    "brave_feature_util.cc",
     "brave_relaunch_utils.cc",
     "brave_shields_content_settings.cc",
     "brave_shields_content_settings.h",
